+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 * In general follow (https://docs.npmjs.com/getting-started/semantic-versioning) versioning.
 
 ## <next>
-* New `info` cell message type and action for setting multiple messages.
+* New `info` cell message type and action for setting multiple messages
+* Update column settings component to utilize `react-select-order-list`
 
 ## 3.2.0
 * Add checkbox to select all columns in column settings